more error fixing
