+++ apps/web/src/app/(content)/(info)/contact/page.tsx
@@ -1,10 +1,11 @@
 import type { Metadata } from 'next';
 import { allPages } from 'contentlayer/generated';
 import { PageCommon } from '@/app/(content)/(info)/page-common';
-import { descriptionHelper } from '@/app/(content)/blog/post-components';
+import { descriptionHelper } from '@/lib/description-helper';
 
 const data = allPages.find((page) => page.url === '/contact');
-const description = descriptionHelper(data?.body.raw, data?.url, true);
+/* eslint-disable @typescript-eslint/no-unnecessary-condition -- testing */
+const description = descriptionHelper(data?.body?.raw, data?.url, true);
 
 export const metadata: Metadata = {
   title: data?.title,
