+++ apps/web/src/app/(content)/blog/[id]/[slug]/page.tsx
@@ -5,11 +5,11 @@ import { compareDesc } from 'date-fns';
 import { useMDXComponent } from 'next-contentlayer2/hooks';
 import { allPosts } from 'contentlayer/generated';
 import ReadingBar from '@/components/reading-bar';
+import { descriptionHelper } from '@/lib/description-helper';
 import { PostHeader2 } from '@/app/(content)/blog/post-header-2';
 import type { FeaturedImageR1 } from '@/lib/image-process';
 import { APP_URL } from '@/lib/constants';
 import BlogImageBlurServer from '../../img-blur-server';
-import { descriptionHelper } from '../../post-components';
 import { catTagData } from '../../cat-tag-roller';
 import CommentsComponent from './comments';
 import { Paragraph } from './paragraph';
