+++ apps/web/src/app/(content)/tags/[slug]/page.tsx
@@ -3,7 +3,7 @@ import type { Metadata, ResolvingMetadata } from 'next';
 import { compareDesc } from 'date-fns';
 import { allTags, allPosts } from 'contentlayer/generated';
 import { MiniLayout } from '@/components/cat-tag-common';
-import { descriptionHelper } from '../../blog/post-components';
+import { descriptionHelper } from '@/lib/description-helper';
 
 export const dynamicParams = false;
