+++ apps/web/sst-env.d.ts
@@ -1,4 +1,3 @@
-/* tslint:disable */
 /* eslint-disable */
 /// <reference path="../../sst-env.d.ts" />
 export {};
